Extracted Operation and Element classes (Calculator)
