+++ resources/krow/zeale/guis/calculator/Equation.java
@@ -3,7 +3,6 @@ package krow.zeale.guis.calculator;
 import java.util.ArrayList;
 import java.util.Collection;
 
-import krow.zeale.guis.calculator.Parser.Element;
 import krow.zeale.guis.calculator.exceptions.EmptyEquationException;
 import krow.zeale.guis.calculator.exceptions.UnmatchedParenthesisException;
 
@@ -34,14 +33,14 @@ class Equation extends ArrayList<Object> {
 		throw new UnsupportedOperationException();
 	}
 
-	public void start(Parser.Element element) {
+	public void start(Element element) {
 		if (started)
 			throw new UnsupportedOperationException();
 		super.add(element);
 		started = true;
 	}
 
-	public void add(Parser.Operation operation, Parser.Element element) {
+	public void add(Operation operation, Element element) {
 		if (!started)
 			throw new UnsupportedOperationException();
 		super.add(operation);
@@ -52,21 +51,21 @@ class Equation extends ArrayList<Object> {
 
 		for (byte precedence = 3; precedence > -1; precedence--)
 			for (int i = 2; i < size() && i > 0; i += 2)
-				if (((Parser.Operation) get(--i)).getPrecedence() == precedence) {
-					popin(new Element.Number(((Parser.Operation) remove(i))
-							.evaluate(((Parser.Element) remove(--i)).evaluate(), ((Parser.Element) remove(i)).evaluate())),
+				if (((Operation) get(--i)).getPrecedence() == precedence) {
+					popin(new Element.Number(((Operation) remove(i))
+							.evaluate(((Element) remove(--i)).evaluate(), ((Element) remove(i)).evaluate())),
 							i);
 					i -= 2;
 				} else
 					i++;
-		return ((Parser.Element) get(0)).evaluate();
+		return ((Element) get(0)).evaluate();
 	}
 
-	private void popin(Parser.Element element, int location) {
+	private void popin(Element element, int location) {
 		super.add(location, element);
 	}
 
-	private void popin(Parser.Operation operation, int location) {
+	private void popin(Operation operation, int location) {
 		super.add(location, operation);
 	}
 }
